requirements.txt
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
grpcio==1.51.3
idna==3.4
numpy==1.24.2
pandas==1.5.3
protobuf==3.20.3
pycodestyle==2.10.0
pydantic==1.10.6
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b50
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
